Rename files and create types file
