+++ typedData/typedData.go
@@ -1,4 +1,4 @@
-package typed
+package typedData
 
 import (
 	"bytes"
